# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PAD Lazuardy API",
        "description": "API documentation for PAD Lazuardy backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Local API Server"
        }
    ],
    "paths": {
...
        "/api/admin/tutor-salary": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List tutor salaries",
                "operationId": "b3dab2fba01f587a809ce61a6d5a0a53",
                "responses": {
                    "200": {
                        "description": "Salaries",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/tutor-salary/{userId}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Tutor salary detail",
                "operationId": "280d27b90051071905ec478f884f5495",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/tutor-salary/{userId}/confirm": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Confirm tutor salary payment",
                "operationId": "01e0eee81223ecd87cef0ea408ec1f29",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Confirmed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/tutor-salary/confirm-batch": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Confirm batch salary payments",
                "operationId": "32b4911c0ba1c3346ec708b920bc9ab9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Batch confirmed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/tutor-salary/{userId}/confirm-with-invoice": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Confirm salary payment with invoice",
                "operationId": "774a3dec54f8247917133ce00cb6259a",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Confirmed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/tutor-salary/pending-payment": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List pending salary payments",
                "operationId": "b358d58bee5f5012ae5a1dc12fca4661",
                "responses": {
                    "200": {
                        "description": "Pending",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/tutor/verification-pending": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List tutor verification pending for salary",
                "operationId": "6ea9d86742796eac98dc56ca4cfd4615",
                "responses": {
                    "200": {
                        "description": "Pending verification",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/tutor-salary/{userId}/history": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Tutor salary history",
                "operationId": "190fb5d9945907b7bb6e9763c8ea652e",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "History",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/verify/tutor": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "List tutors pending verification",
                "operationId": "d3697ee83e20c2289d13aa294f7314a8",
                "responses": {
                    "200": {
                        "description": "Pending tutors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/verify/tutor/approve": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Approve tutor verification",
                "operationId": "daa00890d8e3f24722b2b2238ab54cd6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Approved",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/verify/tutor/reject": {
            "patch": {
                "tags": [
                    "Admin"
                ],
                "summary": "Reject tutor verification",
                "operationId": "4988d58de3625eec4d1e11dddcb0b013",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "reason": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rejected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StandardSuccess"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "StandardSuccess": {
                "required": [
                    "status",
                    "message",
                    "data"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "OK"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Use Bearer token from login/register responses",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}